requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.25.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0
websockets==12.0
python-multipart==0.0.9
prometheus-fastapi-instrumentator==6.1.0

# Ingestion/Parsing
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.5
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0

# Retrieval and ML
faiss-cpu==1.8.0.post1
scikit-learn==1.5.1
sentence-transformers==3.0.1
rank-bm25==0.2.2

# Providers
openai==1.40.1
cohere==5.5.8
anthropic==0.34.2

# Testing and templates
pytest==8.3.2
Jinja2==3.1.4
chainlit==1.2.0
aiohttp==3.9.5
python-docx==1.1.2


